remove image plugin
