+++ components/common/CharmEditor/CharmEditor.tsx
@@ -38,6 +38,7 @@ import type { ConnectionEvent } from './components/fiduswriter/ws';
 import { File } from './components/file/File';
 import FloatingMenu from './components/floatingMenu/FloatingMenu';
 import * as iframe from './components/iframe';
+import ResizableImage from './components/image/ResizableImage';
 import { InlineCommentThread } from './components/inlineComment/components/InlineCommentThread';
 import { InlineDatabase } from './components/inlineDatabase/components/InlineDatabase';
 import InlineCommandPalette from './components/inlinePalette/components/InlineCommandPalette';
@@ -51,7 +52,6 @@ import type { CharmNodeViewProps } from './components/nodeView/nodeView';
 import ResizablePDF from './components/pdf/ResizablePDF';
 import { PollNodeView } from './components/poll/PollComponent';
 import Quote from './components/quote/components/Quote';
-import ResizableImage from './components/ResizableImage';
 import RowActionsMenu from './components/rowActions/RowActionsMenu';
 import { SuggestionsPopup } from './components/suggestions/SuggestionPopup';
 import { TableOfContents } from './components/tableOfContents/TableOfContents';
