+++ components/common/CharmEditor/components/image/ResizableImage.tsx
@@ -8,20 +8,17 @@ import Alert from '@mui/material/Alert';
 import type { HTMLAttributes } from 'react';
 import { memo, useEffect, useRef, useState } from 'react';
 
-import type { RawSpecs } from 'components/common/CharmEditor/components/@bangle.dev/core/specRegistry';
 import ImageSelector from 'components/common/ImageSelector/ImageSelector';
 import LoadingComponent from 'components/common/LoadingComponent';
 import { uploadToS3 } from 'lib/aws/uploadToS3Browser';
 import { MAX_IMAGE_WIDTH, MIN_IMAGE_WIDTH } from 'lib/prosemirror/plugins/image/constants';
 import { replaceS3Domain } from 'lib/utils/url';
 
-import { enableDragAndDrop } from '../utils';
-
-import { getFileBinary } from './@bangle.dev/base-components/image';
-import { spec as suggestTooltipSpec } from './@bangle.dev/tooltip/suggestTooltipSpec';
-import BlockAligner from './BlockAligner';
-import type { CharmNodeViewProps } from './nodeView/nodeView';
-import Resizable from './Resizable/Resizable';
+import { enableDragAndDrop } from '../../utils';
+import { getFileBinary } from '../@bangle.dev/base-components/image';
+import BlockAligner from '../BlockAligner';
+import type { CharmNodeViewProps } from '../nodeView/nodeView';
+import Resizable from '../Resizable/Resizable';
 
 const StyledEmptyImageContainer = styled(Box)`
   display: flex;
@@ -76,59 +73,6 @@ const StyledImage = styled.img`
     cursor: initial;
   }
 `;
-
-function imageSpec(): RawSpecs {
-  return {
-    type: 'node',
-    name: 'image',
-    schema: {
-      inline: false,
-      attrs: {
-        caption: {
-          default: null
-        },
-        src: {
-          default: null
-        },
-        alt: {
-          default: null
-        },
-        size: {
-          // Making sure default size is middle of max and min range
-          default: (MIN_IMAGE_WIDTH + MAX_IMAGE_WIDTH) / 2
-        },
-        track: {
-          default: []
-        }
-      },
-      group: 'block',
-      draggable: false,
-      parseDOM: [
-        {
-          tag: 'img[src]',
-          getAttrs: (dom: any) => ({
-            src: dom.getAttribute('src'),
-            alt: dom.getAttribute('alt')
-          })
-        }
-      ] as any,
-      toDOM: ((node: Node) => {
-        return ['img', node.attrs];
-      }) as any
-    },
-    markdown: {
-      toMarkdown: (state, node) => {
-        const { src } = node.attrs;
-
-        if (src) {
-          const toWrite = `[](${src})`;
-          state.text(toWrite, false);
-          state.ensureNewLine();
-        }
-      }
-    }
-  };
-}
 
 function ResizableImage({
   readOnly = false,
@@ -241,13 +185,6 @@ function ResizableImage({
       )}
     </Box>
   );
-}
-
-export function spec() {
-  // this is a dummy marker to let us know to show the image selector
-  const tooltipSpec = suggestTooltipSpec({ markName: 'tooltip-marker', trigger: 'image', excludes: '_' });
-  tooltipSpec.schema.inclusive = false;
-  return [tooltipSpec, imageSpec()];
 }
 
 export default memo(ResizableImage);
