+++ components/common/CharmEditor/specRegistry.ts
@@ -19,6 +19,7 @@ import { hardBreakSpec } from './components/hardBreak';
 import * as heading from './components/heading';
 import * as horizontalRule from './components/horizontalRule';
 import * as iframe from './components/iframe';
+import * as image from './components/image/imageSpecs';
 import * as inlineComment from './components/inlineComment/inlineComment.specs';
 import * as inlineDatabase from './components/inlineDatabase';
 import { spec as inlinePaletteSpecs } from './components/inlinePalette/inlinePaletteSpec';
@@ -38,7 +39,6 @@ import { spec as paragraphSpec } from './components/paragraph/paragraph';
 import { spec as pdfSpec } from './components/pdf/pdf';
 import * as poll from './components/poll/pollSpec';
 import * as quote from './components/quote/quote';
-import * as image from './components/ResizableImage';
 import * as tabIndent from './components/tabIndent';
 import { spec as tableSpec } from './components/table/table.schema';
 import { spec as tableOfContentSpec } from './components/tableOfContents/tableOfContents.specs';
