+++ tsconfig.json
@@ -58,4 +58,5 @@
     }
   },
   "exclude": ["node_modules"],
+  "include": ["**/*.ts", "**/*.tsx"]
 }
